refactor travel window, jesus what a mess that was
