+++ GWToolboxdll/Windows/TravelWindowConstants.h
@@ -1,0 +1,889 @@
+#pragma once
+
+struct OutpostAlias {
+    OutpostAlias(const GW::Constants::MapID m, const GW::Constants::District d = GW::Constants::District::Current, const uint8_t n = 0)
+        : map_id(m)
+        , district(d)
+        , district_number(n) { }
+
+    GW::Constants::MapID map_id = GW::Constants::MapID::None;
+    GW::Constants::District district = GW::Constants::District::Current;
+    uint8_t district_number = 0;
+};
+
+struct DistrictAlias {
+    DistrictAlias(const GW::Constants::District d, const uint8_t n = 0)
+        : district(d)
+        , district_number(n) { }
+
+    GW::Constants::District district = GW::Constants::District::Current;
+    uint8_t district_number = 0;
+};
+
+// List of shorthand outpost names. This is checked for an exact match first.
+const std::map<const std::string, const OutpostAlias> shorthand_outpost_names = {
+    {"bestarea", {GW::Constants::MapID::The_Deep}},
+    {"toa", {GW::Constants::MapID::Temple_of_the_Ages}},
+    {"doa", {GW::Constants::MapID::Domain_of_Anguish}},
+    {"goa", {GW::Constants::MapID::Domain_of_Anguish}},
+    {"tdp", {GW::Constants::MapID::Domain_of_Anguish}},
+    {"eee", {GW::Constants::MapID::Embark_Beach, GW::Constants::District::EuropeEnglish}},
+    {"gtob", {GW::Constants::MapID::Great_Temple_of_Balthazar_outpost}},
+    {"la", {GW::Constants::MapID::Lions_Arch_outpost}},
+    {"ac", {GW::Constants::MapID::Ascalon_City_outpost}},
+    {"eotn", {GW::Constants::MapID::Eye_of_the_North_outpost}},
+    {"kc", {GW::Constants::MapID::Kaineng_Center_outpost}},
+    {"hzh", {GW::Constants::MapID::House_zu_Heltzer_outpost}},
+    {"ctc", {GW::Constants::MapID::Central_Transfer_Chamber_outpost}},
+    {"topk", {GW::Constants::MapID::Tomb_of_the_Primeval_Kings}},
+    {"ra", {GW::Constants::MapID::Random_Arenas_outpost}},
+    {"ha", {GW::Constants::MapID::Heroes_Ascent_outpost}},
+    {"ra", {GW::Constants::MapID::Random_Arenas_outpost}},
+    {"fa", {GW::Constants::MapID::Fort_Aspenwood_Kurzick_outpost}},
+    {"jq", {GW::Constants::MapID::The_Jade_Quarry_Kurzick_outpost}}
+};
+
+// List of shorthand district names. This is checked for an exact match.
+const std::map<const std::string, const DistrictAlias> shorthand_district_names =
+{
+    {"ae", {GW::Constants::District::American}},
+    {"int", {GW::Constants::District::International}},
+    {"ee", {GW::Constants::District::EuropeEnglish}},
+    {"eg", {GW::Constants::District::EuropeGerman}},
+    {"de", {GW::Constants::District::EuropeGerman}},
+    {"dd", {GW::Constants::District::EuropeGerman}},
+    {"fr", {GW::Constants::District::EuropeFrench}},
+    {"it", {GW::Constants::District::EuropeItalian}},
+    {"es", {GW::Constants::District::EuropeSpanish}},
+    {"pl", {GW::Constants::District::EuropePolish}},
+    {"ru", {GW::Constants::District::EuropeRussian}},
+    {"kr", {GW::Constants::District::AsiaKorean}},
+    {"cn", {GW::Constants::District::AsiaChinese}},
+    {"jp", {GW::Constants::District::AsiaJapanese}},
+};
+
+constexpr std::array presearing_map_ids = {
+    GW::Constants::MapID::Ashford_Abbey_outpost,
+    GW::Constants::MapID::Ascalon_City_pre_searing,
+    GW::Constants::MapID::Foibles_Fair_outpost,
+    GW::Constants::MapID::Fort_Ranik_pre_Searing_outpost,
+    GW::Constants::MapID::The_Barradin_Estate_outpost
+};
+
+constexpr std::array presearing_map_names = {
+    "ashford abbey",
+    "ascalon city",
+    "foibles fair",
+    "fort ranik",
+    "barradin estate"
+};
+
+constexpr std::array searchable_map_ids = {
+    GW::Constants::MapID::Abaddons_Gate,
+    GW::Constants::MapID::Abaddons_Mouth,
+    GW::Constants::MapID::Altrumm_Ruins,
+    GW::Constants::MapID::Amatz_Basin,
+    GW::Constants::MapID::The_Amnoon_Oasis_outpost,
+    GW::Constants::MapID::Arborstone_outpost_mission,
+    GW::Constants::MapID::Ascalon_City_outpost,
+    GW::Constants::MapID::Aspenwood_Gate_Kurzick_outpost,
+    GW::Constants::MapID::Aspenwood_Gate_Luxon_outpost,
+    GW::Constants::MapID::The_Astralarium_outpost,
+    GW::Constants::MapID::Augury_Rock_outpost,
+    GW::Constants::MapID::The_Aurios_Mines,
+    GW::Constants::MapID::Aurora_Glade,
+    GW::Constants::MapID::Bai_Paasu_Reach_outpost,
+    GW::Constants::MapID::Basalt_Grotto_outpost,
+    GW::Constants::MapID::Beacons_Perch_outpost,
+    GW::Constants::MapID::Beetletun_outpost,
+    GW::Constants::MapID::Beknur_Harbor,
+    GW::Constants::MapID::Bergen_Hot_Springs_outpost,
+    GW::Constants::MapID::Blacktide_Den,
+    GW::Constants::MapID::Bloodstone_Fen,
+    GW::Constants::MapID::Bone_Palace_outpost,
+    GW::Constants::MapID::Boreal_Station_outpost,
+    GW::Constants::MapID::Boreas_Seabed_outpost_mission,
+    GW::Constants::MapID::Borlis_Pass,
+    GW::Constants::MapID::Brauer_Academy_outpost,
+    GW::Constants::MapID::Breaker_Hollow_outpost,
+    GW::Constants::MapID::Camp_Hojanu_outpost,
+    GW::Constants::MapID::Camp_Rankor_outpost,
+    GW::Constants::MapID::Cavalon_outpost,
+    GW::Constants::MapID::Central_Transfer_Chamber_outpost,
+    GW::Constants::MapID::Chahbek_Village,
+    GW::Constants::MapID::Champions_Dawn_outpost,
+    GW::Constants::MapID::Chantry_of_Secrets_outpost,
+    GW::Constants::MapID::Codex_Arena_outpost,
+    GW::Constants::MapID::Consulate_Docks,
+    GW::Constants::MapID::Copperhammer_Mines_outpost,
+    GW::Constants::MapID::DAlessio_Seaboard,
+    GW::Constants::MapID::Dajkah_Inlet,
+    GW::Constants::MapID::Dasha_Vestibule,
+    GW::Constants::MapID::The_Deep,
+    GW::Constants::MapID::Deldrimor_War_Camp_outpost,
+    GW::Constants::MapID::Destinys_Gorge_outpost,
+    GW::Constants::MapID::Divinity_Coast,
+    GW::Constants::MapID::Doomlore_Shrine_outpost,
+    GW::Constants::MapID::The_Dragons_Lair,
+    GW::Constants::MapID::Dragons_Throat,
+    GW::Constants::MapID::Droknars_Forge_outpost,
+    GW::Constants::MapID::Druids_Overlook_outpost,
+    GW::Constants::MapID::Dunes_of_Despair,
+    GW::Constants::MapID::Durheim_Archives_outpost,
+    GW::Constants::MapID::Dzagonur_Bastion,
+    GW::Constants::MapID::Elona_Reach,
+    GW::Constants::MapID::Embark_Beach,
+    GW::Constants::MapID::Ember_Light_Camp_outpost,
+    GW::Constants::MapID::Eredon_Terrace_outpost,
+    GW::Constants::MapID::The_Eternal_Grove_outpost_mission,
+    GW::Constants::MapID::Eye_of_the_North_outpost,
+    GW::Constants::MapID::Fishermens_Haven_outpost,
+    GW::Constants::MapID::Fort_Aspenwood_Kurzick_outpost,
+    GW::Constants::MapID::Fort_Aspenwood_Kurzick_outpost,
+    GW::Constants::MapID::Fort_Aspenwood_Luxon_outpost,
+    GW::Constants::MapID::Fort_Aspenwood_Luxon_outpost,
+    GW::Constants::MapID::Fort_Ranik,
+    GW::Constants::MapID::Frontier_Gate_outpost,
+    GW::Constants::MapID::The_Frost_Gate,
+    GW::Constants::MapID::Gadds_Encampment_outpost,
+    GW::Constants::MapID::Domain_of_Anguish,
+    GW::Constants::MapID::Gate_of_Desolation,
+    GW::Constants::MapID::Gate_of_Fear_outpost,
+    GW::Constants::MapID::Gate_of_Madness,
+    GW::Constants::MapID::Gate_of_Pain,
+    GW::Constants::MapID::Gate_of_Secrets_outpost,
+    GW::Constants::MapID::Gate_of_the_Nightfallen_Lands_outpost,
+    GW::Constants::MapID::Gate_of_Torment_outpost,
+    GW::Constants::MapID::Gates_of_Kryta,
+    GW::Constants::MapID::Grand_Court_of_Sebelkeh,
+    GW::Constants::MapID::The_Granite_Citadel_outpost,
+    GW::Constants::MapID::The_Great_Northern_Wall,
+    GW::Constants::MapID::Great_Temple_of_Balthazar_outpost,
+    GW::Constants::MapID::Grendich_Courthouse_outpost,
+    GW::Constants::MapID::Gunnars_Hold_outpost,
+    GW::Constants::MapID::Gyala_Hatchery_outpost_mission,
+    GW::Constants::MapID::Harvest_Temple_outpost,
+    GW::Constants::MapID::Hells_Precipice,
+    GW::Constants::MapID::Henge_of_Denravi_outpost,
+    GW::Constants::MapID::Heroes_Ascent_outpost,
+    GW::Constants::MapID::Heroes_Audience_outpost,
+    GW::Constants::MapID::Honur_Hill_outpost,
+    GW::Constants::MapID::House_zu_Heltzer_outpost,
+    GW::Constants::MapID::Ice_Caves_of_Sorrow,
+    GW::Constants::MapID::Ice_Tooth_Cave_outpost,
+    GW::Constants::MapID::Imperial_Sanctum_outpost_mission,
+    GW::Constants::MapID::Iron_Mines_of_Moladune,
+    GW::Constants::MapID::Jade_Flats_Kurzick_outpost,
+    GW::Constants::MapID::Jade_Flats_Luxon_outpost,
+    GW::Constants::MapID::The_Jade_Quarry_Kurzick_outpost,
+    GW::Constants::MapID::The_Jade_Quarry_Kurzick_outpost,
+    GW::Constants::MapID::The_Jade_Quarry_Luxon_outpost,
+    GW::Constants::MapID::The_Jade_Quarry_Luxon_outpost,
+    GW::Constants::MapID::Jennurs_Horde,
+    GW::Constants::MapID::Jokanur_Diggings,
+    GW::Constants::MapID::Kaineng_Center_outpost,
+    GW::Constants::MapID::Kamadan_Jewel_of_Istan_outpost,
+    GW::Constants::MapID::The_Kodash_Bazaar_outpost,
+    GW::Constants::MapID::Kodlonu_Hamlet_outpost,
+    GW::Constants::MapID::Kodonur_Crossroads,
+    GW::Constants::MapID::Lair_of_the_Forgotten_outpost,
+    GW::Constants::MapID::Leviathan_Pits_outpost,
+    GW::Constants::MapID::Lions_Arch_outpost,
+    GW::Constants::MapID::Longeyes_Ledge_outpost,
+    GW::Constants::MapID::Lutgardis_Conservatory_outpost,
+    GW::Constants::MapID::Maatu_Keep_outpost,
+    GW::Constants::MapID::Maguuma_Stade_outpost,
+    GW::Constants::MapID::Marhans_Grotto_outpost,
+    GW::Constants::MapID::The_Marketplace_outpost,
+    GW::Constants::MapID::Mihanu_Township_outpost,
+    GW::Constants::MapID::Minister_Chos_Estate_outpost_mission,
+    GW::Constants::MapID::Moddok_Crevice,
+    GW::Constants::MapID::The_Mouth_of_Torment_outpost,
+    GW::Constants::MapID::Nahpui_Quarter_outpost_mission,
+    GW::Constants::MapID::Nolani_Academy,
+    GW::Constants::MapID::Nundu_Bay,
+    GW::Constants::MapID::Olafstead_outpost,
+    GW::Constants::MapID::Piken_Square_outpost,
+    GW::Constants::MapID::Pogahn_Passage,
+    GW::Constants::MapID::Port_Sledge_outpost,
+    GW::Constants::MapID::Quarrel_Falls_outpost,
+    GW::Constants::MapID::Raisu_Palace_outpost_mission,
+    GW::Constants::MapID::Ran_Musu_Gardens_outpost,
+    GW::Constants::MapID::Random_Arenas_outpost,
+    GW::Constants::MapID::Rata_Sum_outpost,
+    GW::Constants::MapID::Remains_of_Sahlahja,
+    GW::Constants::MapID::Rilohn_Refuge,
+    GW::Constants::MapID::Ring_of_Fire,
+    GW::Constants::MapID::Riverside_Province,
+    GW::Constants::MapID::Ruins_of_Morah,
+    GW::Constants::MapID::Ruins_of_Morah,
+    GW::Constants::MapID::Ruins_of_Surmia,
+    GW::Constants::MapID::Ruins_of_Surmia,
+    GW::Constants::MapID::Saint_Anjekas_Shrine_outpost,
+    GW::Constants::MapID::Sanctum_Cay,
+    GW::Constants::MapID::Sardelac_Sanitarium_outpost,
+    GW::Constants::MapID::Seafarers_Rest_outpost,
+    GW::Constants::MapID::Seekers_Passage_outpost,
+    GW::Constants::MapID::Seitung_Harbor_outpost,
+    GW::Constants::MapID::Senjis_Corner_outpost,
+    GW::Constants::MapID::Serenity_Temple_outpost,
+    GW::Constants::MapID::The_Shadow_Nexus,
+    GW::Constants::MapID::Shing_Jea_Arena,
+    GW::Constants::MapID::Shing_Jea_Monastery_outpost,
+    GW::Constants::MapID::Sifhalla_outpost,
+    GW::Constants::MapID::Sunjiang_District_outpost_mission,
+    GW::Constants::MapID::Sunspear_Arena,
+    GW::Constants::MapID::Sunspear_Great_Hall_outpost,
+    GW::Constants::MapID::Sunspear_Sanctuary_outpost,
+    GW::Constants::MapID::Tahnnakai_Temple_outpost_mission,
+    GW::Constants::MapID::Tanglewood_Copse_outpost,
+    GW::Constants::MapID::Tarnished_Haven_outpost,
+    GW::Constants::MapID::Temple_of_the_Ages,
+    GW::Constants::MapID::Thirsty_River,
+    GW::Constants::MapID::Thunderhead_Keep,
+    GW::Constants::MapID::Tihark_Orchard,
+    GW::Constants::MapID::Tomb_of_the_Primeval_Kings,
+    GW::Constants::MapID::Tsumei_Village_outpost,
+    GW::Constants::MapID::Umbral_Grotto_outpost,
+    GW::Constants::MapID::Unwaking_Waters_Kurzick_outpost,
+    GW::Constants::MapID::Unwaking_Waters_Luxon_outpost,
+    GW::Constants::MapID::Urgozs_Warren,
+    GW::Constants::MapID::Vasburg_Armory_outpost,
+    GW::Constants::MapID::Venta_Cemetery,
+    GW::Constants::MapID::Ventaris_Refuge_outpost,
+    GW::Constants::MapID::Vizunah_Square_Foreign_Quarter_outpost,
+    GW::Constants::MapID::Vizunah_Square_Local_Quarter_outpost,
+    GW::Constants::MapID::Vloxs_Falls,
+    GW::Constants::MapID::Wehhan_Terraces_outpost,
+    GW::Constants::MapID::The_Wilds,
+    GW::Constants::MapID::Yahnur_Market_outpost,
+    GW::Constants::MapID::Yaks_Bend_outpost,
+    GW::Constants::MapID::Yohlon_Haven_outpost,
+    GW::Constants::MapID::Zaishen_Challenge_outpost,
+    GW::Constants::MapID::Zaishen_Elite_outpost,
+    GW::Constants::MapID::Zaishen_Menagerie_outpost,
+    GW::Constants::MapID::Zen_Daijun_outpost_mission,
+    GW::Constants::MapID::Zin_Ku_Corridor_outpost,
+    GW::Constants::MapID::Zos_Shivros_Channel,
+    GW::Constants::MapID::Great_Temple_of_Balthazar_outpost
+};
+
+constexpr std::array dungeon_map_ids{
+    GW::Constants::MapID::Doomlore_Shrine_outpost,
+    GW::Constants::MapID::Doomlore_Shrine_outpost,
+    GW::Constants::MapID::Doomlore_Shrine_outpost,
+    GW::Constants::MapID::Doomlore_Shrine_outpost,
+    GW::Constants::MapID::Longeyes_Ledge_outpost,
+    GW::Constants::MapID::Longeyes_Ledge_outpost,
+    GW::Constants::MapID::Sifhalla_outpost,
+    GW::Constants::MapID::Sifhalla_outpost,
+    GW::Constants::MapID::Olafstead_outpost,
+    GW::Constants::MapID::Umbral_Grotto_outpost,
+    GW::Constants::MapID::Gadds_Encampment_outpost,
+    GW::Constants::MapID::Deldrimor_War_Camp_outpost
+};
+
+constexpr std::array searchable_dungeon_names{
+    "catacombs of kathandrax",
+    "kathandrax",
+    "rragars menagerie",
+    "cathedral of flames",
+    "ooze pit",
+    "darkrime delves",
+    "frostmaws burrows",
+    "sepulchre of dragrimmar",
+    "ravens point",
+    "vloxen excavations",
+    "bogroot growths",
+    "sorrows furnace"
+};
+
+constexpr std::array searchable_map_names{
+    "abaddons gate",
+    "abaddons mouth",
+    "altrumm ruins",
+    "amatz basin",
+    "amnoon oasis",
+    "arborstone",
+    "ascalon city",
+    "aspenwood gate kurzick",
+    "aspenwood gate luxon",
+    "astralarium",
+    "augury rock",
+    "aurios mines",
+    "aurora glade",
+    "bai paasu reach",
+    "basalt grotto",
+    "beacons perch",
+    "beetletun",
+    "beknur harbor",
+    "bergen hot springs",
+    "blacktide den",
+    "bloodstone fen",
+    "bone palace",
+    "boreal station",
+    "boreas seabed",
+    "borlis pass",
+    "brauer academy",
+    "breaker hollow",
+    "camp hojanu",
+    "camp rankor",
+    "cavalon",
+    "central transfer chamber",
+    "chahbek village",
+    "champions dawn",
+    "chantry of secrets",
+    "codex arena",
+    "consulate docks",
+    "copperhammer mines",
+    "dalessio seaboard",
+    "dajkah inlet",
+    "dasha vestibule",
+    "deep",
+    "deldrimor war camp",
+    "destinys gorge",
+    "divinity coast",
+    "doomlore shrine",
+    "dragons lair",
+    "dragons throat",
+    "droknars forge",
+    "druids overlook",
+    "dunes of despair",
+    "durheim archives",
+    "dzagonur bastion",
+    "elona reach",
+    "embark beach",
+    "ember light camp",
+    "eredon terrace",
+    "eternal grove",
+    "eye of the north",
+    "fishermens haven",
+    "fort aspenwood kurzick",
+    "fa kurzick",
+    "fort aspenwood luxon",
+    "fa luxon",
+    "fort ranik",
+    "frontier gate",
+    "frost gate",
+    "gadds encampment",
+    "domain of anguish",
+    "gate of desolation",
+    "gate of fear",
+    "gate of madness",
+    "gate of pain",
+    "gate of secrets",
+    "gate of the nightfallen lands",
+    "gate of torment",
+    "gates of kryta",
+    "grand court of sebelkeh",
+    "granite citadel",
+    "great northern wall",
+    "great temple of balthazar",
+    "grendich courthouse",
+    "gunnars hold",
+    "gyala hatchery",
+    "harvest temple",
+    "hells precipice",
+    "henge of denravi",
+    "heroes ascent",
+    "heroes audience",
+    "honur hill",
+    "house zu heltzer",
+    "ice caves of sorrow",
+    "ice tooth cave",
+    "imperial sanctum",
+    "iron mines of moladune",
+    "jade flats kurzick",
+    "jade flats luxon",
+    "jade quarry kurzick",
+    "jq kurzick",
+    "jade quarry luxon",
+    "jq luxon",
+    "jennurs horde",
+    "jokanur diggings",
+    "kaineng center",
+    "kamadan jewel of istan",
+    "kodash bazaar",
+    "kodlonu hamlet",
+    "kodonur crossroads",
+    "lair of the forgotten",
+    "leviathan pits",
+    "lions arch",
+    "longeyes ledge",
+    "lutgardis conservatory",
+    "maatu keep",
+    "maguuma stade",
+    "marhans grotto",
+    "marketplace",
+    "mihanu township",
+    "minister chos estate",
+    "moddok crevice",
+    "mouth of torment",
+    "nahpui quarter",
+    "nolani academy",
+    "nundu bay",
+    "olafstead",
+    "piken square",
+    "pogahn passage",
+    "port sledge",
+    "quarrel falls",
+    "raisu palace",
+    "ran musu gardens",
+    "random arenas",
+    "rata sum",
+    "remains of sahlahja",
+    "rilohn refuge",
+    "ring of fire",
+    "riverside province",
+    "ruins of morah",
+    "morah",
+    "ruins of surmia",
+    "surmia",
+    "saint anjekas shrine",
+    "sanctum cay",
+    "sardelac sanitarium",
+    "seafarers rest",
+    "seekers passage",
+    "seitung harbor",
+    "senjis corner",
+    "serenity temple",
+    "shadow nexus",
+    "shing jea arena",
+    "shing jea monastery",
+    "sifhalla",
+    "sunjiang district",
+    "sunspear arena",
+    "sunspear great hall",
+    "sunspear sanctuary",
+    "tahnnakai temple",
+    "tanglewood copse",
+    "tarnished haven",
+    "temple of the ages",
+    "thirsty river",
+    "thunderhead keep",
+    "tihark orchard",
+    "tomb of the primeval kings",
+    "tsumei village",
+    "umbral grotto",
+    "unwaking waters kurzick",
+    "unwaking waters luxon",
+    "urgozs warren",
+    "vasburg armory",
+    "venta cemetery",
+    "ventaris refuge",
+    "vizunah square foreign quarter",
+    "vizunah square local quarter",
+    "vloxs falls",
+    "wehhan terraces",
+    "wilds",
+    "yahnur market",
+    "yaks bend",
+    "yohlon haven",
+    "zaishen challenge",
+    "zaishen elite",
+    "zaishen menagerie",
+    "zen daijun",
+    "zin ku corridor",
+    "zos shivros channel",
+    "great temple of balthazar"
+};
+
+constexpr std::array outpost_ids = {
+    GW::Constants::MapID::Abaddons_Gate,
+    GW::Constants::MapID::Abaddons_Mouth,
+    GW::Constants::MapID::Altrumm_Ruins,
+    GW::Constants::MapID::Amatz_Basin,
+    GW::Constants::MapID::The_Amnoon_Oasis_outpost,
+    GW::Constants::MapID::Arborstone_outpost_mission,
+    GW::Constants::MapID::Ascalon_City_outpost,
+    GW::Constants::MapID::Aspenwood_Gate_Kurzick_outpost,
+    GW::Constants::MapID::Aspenwood_Gate_Luxon_outpost,
+    GW::Constants::MapID::The_Astralarium_outpost,
+    GW::Constants::MapID::Augury_Rock_outpost,
+    GW::Constants::MapID::The_Aurios_Mines,
+    GW::Constants::MapID::Aurora_Glade,
+    GW::Constants::MapID::Bai_Paasu_Reach_outpost,
+    GW::Constants::MapID::Basalt_Grotto_outpost,
+    GW::Constants::MapID::Beacons_Perch_outpost,
+    GW::Constants::MapID::Beetletun_outpost,
+    GW::Constants::MapID::Beknur_Harbor,
+    GW::Constants::MapID::Bergen_Hot_Springs_outpost,
+    GW::Constants::MapID::Blacktide_Den,
+    GW::Constants::MapID::Bloodstone_Fen,
+    GW::Constants::MapID::Bone_Palace_outpost,
+    GW::Constants::MapID::Boreal_Station_outpost,
+    GW::Constants::MapID::Boreas_Seabed_outpost_mission,
+    GW::Constants::MapID::Borlis_Pass,
+    GW::Constants::MapID::Brauer_Academy_outpost,
+    GW::Constants::MapID::Breaker_Hollow_outpost,
+    GW::Constants::MapID::Camp_Hojanu_outpost,
+    GW::Constants::MapID::Camp_Rankor_outpost,
+    GW::Constants::MapID::Cavalon_outpost,
+    GW::Constants::MapID::Central_Transfer_Chamber_outpost,
+    GW::Constants::MapID::Chahbek_Village,
+    GW::Constants::MapID::Champions_Dawn_outpost,
+    GW::Constants::MapID::Chantry_of_Secrets_outpost,
+    GW::Constants::MapID::Codex_Arena_outpost,
+    GW::Constants::MapID::Consulate_Docks,
+    GW::Constants::MapID::Copperhammer_Mines_outpost,
+    GW::Constants::MapID::DAlessio_Seaboard,
+    GW::Constants::MapID::Dajkah_Inlet,
+    GW::Constants::MapID::Dasha_Vestibule,
+    GW::Constants::MapID::The_Deep,
+    GW::Constants::MapID::Deldrimor_War_Camp_outpost,
+    GW::Constants::MapID::Destinys_Gorge_outpost,
+    GW::Constants::MapID::Divinity_Coast,
+    GW::Constants::MapID::Doomlore_Shrine_outpost,
+    GW::Constants::MapID::The_Dragons_Lair,
+    GW::Constants::MapID::Dragons_Throat,
+    GW::Constants::MapID::Droknars_Forge_outpost,
+    GW::Constants::MapID::Druids_Overlook_outpost,
+    GW::Constants::MapID::Dunes_of_Despair,
+    GW::Constants::MapID::Durheim_Archives_outpost,
+    GW::Constants::MapID::Dzagonur_Bastion,
+    GW::Constants::MapID::Elona_Reach,
+    GW::Constants::MapID::Embark_Beach,
+    GW::Constants::MapID::Ember_Light_Camp_outpost,
+    GW::Constants::MapID::Eredon_Terrace_outpost,
+    GW::Constants::MapID::The_Eternal_Grove_outpost_mission,
+    GW::Constants::MapID::Eye_of_the_North_outpost,
+    GW::Constants::MapID::Fishermens_Haven_outpost,
+    GW::Constants::MapID::Fort_Aspenwood_Kurzick_outpost,
+    GW::Constants::MapID::Fort_Aspenwood_Luxon_outpost,
+    GW::Constants::MapID::Fort_Ranik,
+    GW::Constants::MapID::Frontier_Gate_outpost,
+    GW::Constants::MapID::The_Frost_Gate,
+    GW::Constants::MapID::Gadds_Encampment_outpost,
+    GW::Constants::MapID::Domain_of_Anguish,
+    GW::Constants::MapID::Gate_of_Desolation,
+    GW::Constants::MapID::Gate_of_Fear_outpost,
+    GW::Constants::MapID::Gate_of_Madness,
+    GW::Constants::MapID::Gate_of_Pain,
+    GW::Constants::MapID::Gate_of_Secrets_outpost,
+    GW::Constants::MapID::Gate_of_the_Nightfallen_Lands_outpost,
+    GW::Constants::MapID::Gate_of_Torment_outpost,
+    GW::Constants::MapID::Gates_of_Kryta,
+    GW::Constants::MapID::Grand_Court_of_Sebelkeh,
+    GW::Constants::MapID::The_Granite_Citadel_outpost,
+    GW::Constants::MapID::The_Great_Northern_Wall,
+    GW::Constants::MapID::Great_Temple_of_Balthazar_outpost,
+    GW::Constants::MapID::Grendich_Courthouse_outpost,
+    GW::Constants::MapID::Gunnars_Hold_outpost,
+    GW::Constants::MapID::Gyala_Hatchery_outpost_mission,
+    GW::Constants::MapID::Harvest_Temple_outpost,
+    GW::Constants::MapID::Hells_Precipice,
+    GW::Constants::MapID::Henge_of_Denravi_outpost,
+    GW::Constants::MapID::Heroes_Ascent_outpost,
+    GW::Constants::MapID::Heroes_Audience_outpost,
+    GW::Constants::MapID::Honur_Hill_outpost,
+    GW::Constants::MapID::House_zu_Heltzer_outpost,
+    GW::Constants::MapID::Ice_Caves_of_Sorrow,
+    GW::Constants::MapID::Ice_Tooth_Cave_outpost,
+    GW::Constants::MapID::Imperial_Sanctum_outpost_mission,
+    GW::Constants::MapID::Iron_Mines_of_Moladune,
+    GW::Constants::MapID::Jade_Flats_Kurzick_outpost,
+    GW::Constants::MapID::Jade_Flats_Luxon_outpost,
+    GW::Constants::MapID::The_Jade_Quarry_Kurzick_outpost,
+    GW::Constants::MapID::The_Jade_Quarry_Luxon_outpost,
+    GW::Constants::MapID::Jennurs_Horde,
+    GW::Constants::MapID::Jokanur_Diggings,
+    GW::Constants::MapID::Kaineng_Center_outpost,
+    GW::Constants::MapID::Kamadan_Jewel_of_Istan_outpost,
+    GW::Constants::MapID::The_Kodash_Bazaar_outpost,
+    GW::Constants::MapID::Kodlonu_Hamlet_outpost,
+    GW::Constants::MapID::Kodonur_Crossroads,
+    GW::Constants::MapID::Lair_of_the_Forgotten_outpost,
+    GW::Constants::MapID::Leviathan_Pits_outpost,
+    GW::Constants::MapID::Lions_Arch_outpost,
+    GW::Constants::MapID::Longeyes_Ledge_outpost,
+    GW::Constants::MapID::Lutgardis_Conservatory_outpost,
+    GW::Constants::MapID::Maatu_Keep_outpost,
+    GW::Constants::MapID::Maguuma_Stade_outpost,
+    GW::Constants::MapID::Marhans_Grotto_outpost,
+    GW::Constants::MapID::The_Marketplace_outpost,
+    GW::Constants::MapID::Mihanu_Township_outpost,
+    GW::Constants::MapID::Minister_Chos_Estate_outpost_mission,
+    GW::Constants::MapID::Moddok_Crevice,
+    GW::Constants::MapID::The_Mouth_of_Torment_outpost,
+    GW::Constants::MapID::Nahpui_Quarter_outpost_mission,
+    GW::Constants::MapID::Nolani_Academy,
+    GW::Constants::MapID::Nundu_Bay,
+    GW::Constants::MapID::Olafstead_outpost,
+    GW::Constants::MapID::Piken_Square_outpost,
+    GW::Constants::MapID::Pogahn_Passage,
+    GW::Constants::MapID::Port_Sledge_outpost,
+    GW::Constants::MapID::Quarrel_Falls_outpost,
+    GW::Constants::MapID::Raisu_Palace_outpost_mission,
+    GW::Constants::MapID::Ran_Musu_Gardens_outpost,
+    GW::Constants::MapID::Random_Arenas_outpost,
+    GW::Constants::MapID::Rata_Sum_outpost,
+    GW::Constants::MapID::Remains_of_Sahlahja,
+    GW::Constants::MapID::Rilohn_Refuge,
+    GW::Constants::MapID::Ring_of_Fire,
+    GW::Constants::MapID::Riverside_Province,
+    GW::Constants::MapID::Ruins_of_Morah,
+    GW::Constants::MapID::Ruins_of_Surmia,
+    GW::Constants::MapID::Saint_Anjekas_Shrine_outpost,
+    GW::Constants::MapID::Sanctum_Cay,
+    GW::Constants::MapID::Sardelac_Sanitarium_outpost,
+    GW::Constants::MapID::Seafarers_Rest_outpost,
+    GW::Constants::MapID::Seekers_Passage_outpost,
+    GW::Constants::MapID::Seitung_Harbor_outpost,
+    GW::Constants::MapID::Senjis_Corner_outpost,
+    GW::Constants::MapID::Serenity_Temple_outpost,
+    GW::Constants::MapID::The_Shadow_Nexus,
+    GW::Constants::MapID::Shing_Jea_Arena,
+    GW::Constants::MapID::Shing_Jea_Monastery_outpost,
+    GW::Constants::MapID::Sifhalla_outpost,
+    GW::Constants::MapID::Sunjiang_District_outpost_mission,
+    GW::Constants::MapID::Sunspear_Arena,
+    GW::Constants::MapID::Sunspear_Great_Hall_outpost,
+    GW::Constants::MapID::Sunspear_Sanctuary_outpost,
+    GW::Constants::MapID::Tahnnakai_Temple_outpost_mission,
+    GW::Constants::MapID::Tanglewood_Copse_outpost,
+    GW::Constants::MapID::Tarnished_Haven_outpost,
+    GW::Constants::MapID::Temple_of_the_Ages,
+    GW::Constants::MapID::Thirsty_River,
+    GW::Constants::MapID::Thunderhead_Keep,
+    GW::Constants::MapID::Tihark_Orchard,
+    GW::Constants::MapID::Tomb_of_the_Primeval_Kings,
+    GW::Constants::MapID::Tsumei_Village_outpost,
+    GW::Constants::MapID::Umbral_Grotto_outpost,
+    GW::Constants::MapID::Unwaking_Waters_Kurzick_outpost,
+    GW::Constants::MapID::Unwaking_Waters_Luxon_outpost,
+    GW::Constants::MapID::Urgozs_Warren,
+    GW::Constants::MapID::Vasburg_Armory_outpost,
+    GW::Constants::MapID::Venta_Cemetery,
+    GW::Constants::MapID::Ventaris_Refuge_outpost,
+    GW::Constants::MapID::Vizunah_Square_Foreign_Quarter_outpost,
+    GW::Constants::MapID::Vizunah_Square_Local_Quarter_outpost,
+    GW::Constants::MapID::Vloxs_Falls,
+    GW::Constants::MapID::Wehhan_Terraces_outpost,
+    GW::Constants::MapID::The_Wilds,
+    GW::Constants::MapID::Yahnur_Market_outpost,
+    GW::Constants::MapID::Yaks_Bend_outpost,
+    GW::Constants::MapID::Yohlon_Haven_outpost,
+    GW::Constants::MapID::Zaishen_Challenge_outpost,
+    GW::Constants::MapID::Zaishen_Elite_outpost,
+    GW::Constants::MapID::Zaishen_Menagerie_outpost,
+    GW::Constants::MapID::Zen_Daijun_outpost_mission,
+    GW::Constants::MapID::Zin_Ku_Corridor_outpost,
+    GW::Constants::MapID::Zos_Shivros_Channel
+};
+
+constexpr std::array outpost_names = {
+    "Abaddon's Gate",
+    "Abaddon's Mouth",
+    "Altrumm Ruins",
+    "Amatz Basin",
+    "Amnoon Oasis, The",
+    "Arborstone",
+    "Ascalon City",
+    "Aspenwood Gate (Kurzick)",
+    "Aspenwood Gate (Luxon)",
+    "Astralarium, The",
+    "Augury Rock",
+    "Aurios Mines, The",
+    "Aurora Glade",
+    "Bai Paasu Reach",
+    "Basalt Grotto",
+    "Beacon's Perch",
+    "Beetletun",
+    "Beknur Harbor",
+    "Bergen Hot Springs",
+    "Blacktide Den",
+    "Bloodstone Fen",
+    "Bone Palace",
+    "Boreal Station",
+    "Boreas Seabed",
+    "Borlis Pass",
+    "Brauer Academy",
+    "Breaker Hollow",
+    "Camp Hojanu",
+    "Camp Rankor",
+    "Cavalon",
+    "Central Transfer Chamber",
+    "Chahbek Village",
+    "Champion's Dawn",
+    "Chantry of Secrets",
+    "Codex Arena",
+    "Consulate Docks",
+    "Copperhammer Mines",
+    "D'Alessio Seaboard",
+    "Dajkah Inlet",
+    "Dasha Vestibule",
+    "Deep, The",
+    "Deldrimor War Camp",
+    "Destiny's Gorge",
+    "Divinity Coast",
+    "Doomlore Shrine",
+    "Dragon's Lair, The",
+    "Dragon's Throat",
+    "Droknar's Forge",
+    "Druid's Overlook",
+    "Dunes of Despair",
+    "Durheim Archives",
+    "Dzagonur Bastion",
+    "Elona Reach",
+    "Embark Beach",
+    "Ember Light Camp",
+    "Eredon Terrace",
+    "Eternal Grove, The",
+    "Eye of the North",
+    "Fishermen's Haven",
+    "Fort Aspenwood (Kurzick)",
+    "Fort Aspenwood (Luxon)",
+    "Fort Ranik",
+    "Frontier Gate",
+    "Frost Gate, The",
+    "Gadd's Encampment",
+    "Gate of Anguish",
+    "Gate of Desolation",
+    "Gate of Fear",
+    "Gate of Madness",
+    "Gate of Pain",
+    "Gate of Secrets",
+    "Gate of the Nightfallen Lands",
+    "Gate of Torment",
+    "Gates of Kryta",
+    "Grand Court of Sebelkeh",
+    "Granite Citadel, The",
+    "Great Northern Wall, The",
+    "Great Temple of Balthazar",
+    "Grendich Courthouse",
+    "Gunnar's Hold",
+    "Gyala Hatchery",
+    "Harvest Temple",
+    "Hell's Precipice",
+    "Henge of Denravi",
+    "Heroes' Ascent",
+    "Heroes' Audience",
+    "Honur Hill",
+    "House zu Heltzer",
+    "Ice Caves of Sorrow",
+    "Ice Tooth Cave",
+    "Imperial Sanctum",
+    "Iron Mines of Moladune",
+    "Jade Flats (Kurzick)",
+    "Jade Flats (Luxon)",
+    "Jade Quarry (Kurzick), The",
+    "Jade Quarry (Luxon), The",
+    "Jennur's Horde",
+    "Jokanur Diggings",
+    "Kaineng Center",
+    "Kamadan, Jewel of Istan",
+    "Kodash Bazaar, The",
+    "Kodlonu Hamlet",
+    "Kodonur Crossroads",
+    "Lair of the Forgotten",
+    "Leviathan Pits",
+    "Lion's Arch",
+    "Longeye's Ledge",
+    "Lutgardis Conservatory",
+    "Maatu Keep",
+    "Maguuma Stade",
+    "Marhan's Grotto",
+    "Marketplace, The",
+    "Mihanu Township",
+    "Minister Cho's Estate",
+    "Moddok Crevice",
+    "Mouth of Torment, The",
+    "Nahpui Quarter",
+    "Nolani Academy",
+    "Nundu Bay",
+    "Olafstead",
+    "Piken Square",
+    "Pogahn Passage",
+    "Port Sledge",
+    "Quarrel Falls",
+    "Raisu Palace",
+    "Ran Musu Gardens",
+    "Random Arenas",
+    "Rata Sum",
+    "Remains of Sahlahja",
+    "Rilohn Refuge",
+    "Ring of Fire",
+    "Riverside Province",
+    "Ruins of Morah",
+    "Ruins of Surmia",
+    "Saint Anjeka's Shrine",
+    "Sanctum Cay",
+    "Sardelac Sanitarium",
+    "Seafarer's Rest",
+    "Seeker's Passage",
+    "Seitung Harbor",
+    "Senji's Corner",
+    "Serenity Temple",
+    "Shadow Nexus, The",
+    "Shing Jea Arena",
+    "Shing Jea Monastery",
+    "Sifhalla",
+    "Sunjiang District",
+    "Sunspear Arena",
+    "Sunspear Great Hall",
+    "Sunspear Sanctuary",
+    "Tahnnakai Temple",
+    "Tanglewood Copse",
+    "Tarnished Haven",
+    "Temple of the Ages",
+    "Thirsty River",
+    "Thunderhead Keep",
+    "Tihark Orchard",
+    "Tomb of the Primeval Kings",
+    "Tsumei Village",
+    "Umbral Grotto",
+    "Unwaking Waters (Kurzick)",
+    "Unwaking Waters (Luxon)",
+    "Urgoz's Warren",
+    "Vasburg Armory",
+    "Venta Cemetery",
+    "Ventari's Refuge",
+    "Vizunah Square (Foreign)",
+    "Vizunah Square (Local)",
+    "Vlox's Falls",
+    "Wehhan Terraces",
+    "Wilds, The",
+    "Yahnur Market",
+    "Yak's Bend",
+    "Yohlon Haven",
+    "Zaishen Challenge",
+    "Zaishen Elite",
+    "Zaishen Menagerie",
+    "Zen Daijun",
+    "Zin Ku Corridor",
+    "Zos Shivros Channel"
+};
+
+constexpr std::array district_words = {
+    "Current District",
+    "International",
+    "American",
+    "American District 1",
+    "Europe English",
+    "Europe French",
+    "Europe German",
+    "Europe Italian",
+    "Europe Spanish",
+    "Europe Polish",
+    "Europe Russian",
+    "Asian Korean",
+    "Asia Chinese",
+    "Asia Japanese",
+};
+
+constexpr std::array district_ids = {
+    GW::Constants::District::Current,
+    GW::Constants::District::International,
+    GW::Constants::District::American,
+    GW::Constants::District::American,
+    GW::Constants::District::EuropeEnglish,
+    GW::Constants::District::EuropeFrench,
+    GW::Constants::District::EuropeGerman,
+    GW::Constants::District::EuropeItalian,
+    GW::Constants::District::EuropeSpanish,
+    GW::Constants::District::EuropePolish,
+    GW::Constants::District::EuropeRussian,
+    GW::Constants::District::AsiaKorean,
+    GW::Constants::District::AsiaChinese,
+    GW::Constants::District::AsiaJapanese,
+};
